Added road camera points, added optional optimization
Struggling with a bug where sometimes it will skip over some of the road camera points. Need to look more into it, it might be nothing.

Added optimization where it will use dot product to estimate whether a target and all of its obstacles are making a difference so it can skip when they're not.