Fixed some bugs and added some functionality to save the gridpoints in a json file.
Unfortunately it's not working as intended right now, I'll have to fix it later.
I also started functionality in a "viewer mode" that can be used to display previously recorded data, but it still has a lot to be done before it works as intended.
Also, it seems the issue with the gridpoints not being shown may be caused by the particle limit on the particle system. More to be seen there, like running the test with the smaller vehicles to see if the issue persists.